Separate SettingsActivity into fragments
